# docs/features/base-vertical/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: url-shortener — base-vertical
  version: 0.1.0
  description: >
    Shorten a URL, follow a short code, list links, read stats.
    Single-visitor service — no authentication surface.
# No auth: single-visitor service (see docs/CONTEXT.md Out of scope).
security: []
paths:
  /api/shorten:
    post:
      summary: Create a short link
      operationId: createLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortenRequest'
      responses:
        '201':
          description: Link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /{code}:
    get:
      summary: Follow a short code
      operationId: followCode
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the original URL
          headers:
            Location:
              schema:
                type: string
        '404':
          description: Unknown code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/links:
    get:
      summary: List links
      operationId: listLinks
      responses:
        '200':
          description: Links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Link'
  /api/stats/{code}:
    get:
      summary: Read link stats
      operationId: getStats
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '404':
          description: Unknown code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    ShortenRequest:
      type: object
      additionalProperties: false
      required: [url]
      properties:
        url:
          type: string
    Link:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
        url:
          type: string
        short_url:
          type: string
        created_at:
          type: integer
        clicks:
          type: integer
    Stats:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
        clicks:
          type: integer
        created_at:
          type: integer
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string